requirements.txt
Flask>=3.0.0
Bootstrap-Flask>=2.3.3
supervisor>=4.2.5
PyYAML>=6.0.1
Werkzeug>=3.0.1
pytest>=7.4.3
typing-extensions>=4.8.0
requests>=2.31.0
python-dateutil>=2.8.2